Added line width and RenderFlags set and clear to RenderNode init messages ("RFSet" & "RFClear", add the index of the flag) Added more functionality to the TContour class (an outline) used then for generating tesselations. Can [moderately] safely shrink and grow the shape now, check for self intersection etc. Added extra line, sphere, box code. Added xy() and xyz() to Type2/Type3 classes to extract parts of the variable. Fixed some memory overwrite bugs in TArray. Added a check for expensive copy/move loops. Added TLGame file to the game library for some misc little classes/functions. Started to add a Screen ray calculation for projection cameras but it's not correct yet. stick to ortho for now :) Started to add a simple tesselator, (cross platform, just makes triangles) but not implemented yet. Added mesh functionality for removing/replacing vertexes and modifying all the relavant polygons. Added flags to mesh. Currently just a flag to specifiy if there is/isn't any alpha's contained in the vertex colours.

+++ Code/TootleRender/TRenderNodeTile.cpp
@@ -57,8 +57,8 @@ Bool TLRender::TRenderNodeTile::Draw(TRenderTarget* pRenderTarget,TRenderNode* p
 
 	float2 ScreenScale2 = Widescreen ? float2( ScreenScale * 1.5f, ScreenScale * 1.f ) : float2( ScreenScale * 1.f, ScreenScale * 1.5f );
 
-	s32 TilesWide = ( (ScreenScale2.x*CameraZ) / TileSize.x) / 2.f;
-	s32 TilesHigh = ( (ScreenScale2.y*CameraZ) / TileSize.y) / 2.f;
+	s32 TilesWide = (s32)(( (ScreenScale2.x*CameraZ) / TileSize.x) / 2.f);
+	s32 TilesHigh = (s32)(( (ScreenScale2.y*CameraZ) / TileSize.y) / 2.f);
 	u32 NodeCount = 0;
 
 	if ( pCamera->IsOrtho() )
